+++ virtualization/axaddrspace/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.15](https://github.com/rcore-os/tgoskits/compare/axaddrspace-v0.5.14...axaddrspace-v0.5.15) - 2026-07-02
+
+### Other
+
+- *(axvm)* redesign guest address layout planning ([#1454](https://github.com/rcore-os/tgoskits/pull/1454))
+
 ## [0.5.14](https://github.com/rcore-os/tgoskits/compare/axaddrspace-v0.5.13...axaddrspace-v0.5.14) - 2026-06-22
 
 ### Other
@@ -69,7 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.1
 
-- Support the new 4-level-ept feature. By default, level 3 ept is used. After enabling this feature, level 4 ept is used.
+- Support the former four-level EPT build option. By default, level 3 EPT is used. After enabling this option, level 4 EPT is used.
 
 ## 0.1.0